+++ azdev/params.py
@@ -53,9 +53,12 @@ def load_arguments(self, _):
         with ArgumentsContext(self, 'verify {}'.format(scope)) as c:
             c.positional('modules', nargs='*', help='Space-separated list of modules to check.')
 
-    with ArgumentsContext(self, 'verify version') as c:
+    with ArgumentsContext(self, 'cli check-versions') as c:
         c.argument('update', action='store_true', help='If provided, the command will update the versions in azure-cli\'s setup.py file.')
         c.argument('pin', action='store_true', help='If provided and used with --update, will pin the module versions in azure-cli\'s setup.py file.')
+
+    with ArgumentsContext(self, 'cli update-setup') as c:
+        c.argument('pin', action='store_true', help='Pin the module versions in azure-cli\'s setup.py file.')
 
     with ArgumentsContext(self, 'linter') as c:
         c.positional('modules', nargs='*', help='Space-separated list of modules or extensions to check.')
@@ -73,9 +76,9 @@ def load_arguments(self, _):
         c.argument('update_index', action='store_true', help='Update the index.json file after publishing is complete.')
 
     with ArgumentsContext(self, 'extension publish') as c:
-        c.argument('storage_account', help='Name of the storage account to publish to.', arg_group='Storage', configured_default='storage_account')
-        c.argument('storage_container', help='Name of the storage container to publish to.', arg_group='Storage', configured_default='storage_container')
-        c.argument('storage_subscription', help='Subscription ID of the storage account.', arg_group='Storage', configured_default='storage_subscription')
+        c.argument('storage_account', help='Name of the storage account to publish to. Environment variable: AZDEV_DEFAULTS_STORAGE_ACCOUNT.', arg_group='Storage', configured_default='storage_account')
+        c.argument('storage_container', help='Name of the storage container to publish to. Environment variable: AZDEV_DEFAULTS_STORAGE_CONTAINER.', arg_group='Storage', configured_default='storage_container')
+        c.argument('storage_subscription', help='Subscription ID of the storage account. Environment variable: AZDEV_DEFAULTS_STORAGE_SUBSCRIPTION.', arg_group='Storage', configured_default='storage_subscription')
 
     for scope in ['extension add', 'extension remove', 'extension build', 'extension publish']:
         with ArgumentsContext(self, scope) as c:
@@ -90,3 +93,25 @@ def load_arguments(self, _):
 
     with ArgumentsContext(self, 'cli check-versions') as c:
         c.positional('modules', nargs='*', help='Space-separated list of modules to check.')
+
+    with ArgumentsContext(self, 'cli create') as c:
+        c.positional('mod_name', help='Name of the module to create.')
+
+    with ArgumentsContext(self, 'extension create') as c:
+        c.positional('ext_name', help='Name of the extension to create.')
+
+    for scope in ['extension create', 'cli create']:
+        with ArgumentsContext(self, scope) as c:
+            c.argument('github_alias', help='Github alias for the individual who will be the code owner for this package.')
+            c.argument('not_preview', action='store_true', help='Do not create template commands under a "Preview" status.')
+            c.argument('required_sdk', help='Name and version of the underlying Azure SDK that is published on PyPI. (ex: azure-mgmt-contoso==0.1.0).', arg_group='SDK')
+            c.argument('local_sdk', help='Path to a locally saved SDK. Use if your SDK is not available on PyPI.', arg_group='SDK')
+            c.argument('client_name', help='Name of the Python SDK client object (ex: ContosoManagementClient).', arg_group='SDK')
+            c.argument('operation_name', help='Name of the principal Python SDK operation class (ex: ContosoOperations).', arg_group='SDK')
+            c.argument('sdk_property', help='The name of the Python variable that describes the main object name in the SDK calls (i.e.: account_name)', arg_group='SDK')
+            c.argument('repo_name', help='Name of the repo the extension will exist in.')
+            c.argument('display_name', arg_group='Help', help='Description to display in help text.')
+            c.argument('display_name_plural', arg_group='Help', help='Description to display in help text when plural.')
+
+    with ArgumentsContext(self, 'cli create') as c:
+        c.ignore('local_sdk')